requirements.txt
streamlitstreamlit==1.24.0
pandas==2.1.1
numpy==1.24.4
statsmodels==0.14.0
matplotlib==3.8.0